Upodated build script ot be the same as in the 0.5.x branch

+++ build.ps1
@@ -1,2 +1,23 @@
-Remove-Item -Recurse -Force build/*
-dotnet publish -c Release -o build
+# Remove build folder & contents if it exists
+if (Test-Path -Path build) {
+    Remove-Item -Recurse -Force build/*
+}
+
+dotnet publish -c Release -o temp
+
+# create build folder if it doesn't exist
+if (!(Test-Path -Path .\build)) {
+    New-Item -ItemType Directory -Path .\build
+}
+
+# copy AdditionalKeys.dll to build folder
+Copy-Item .\temp\AdditionalKeys.dll .\build\AdditionalKeys.dll
+Copy-Item .\temp\AdditionalKeys.pdb .\build\AdditionalKeys.pdb
+
+# remove temp folder
+
+Remove-Item -Recurse -Force temp
+
+# zip the contents of the build folder
+
+Compress-Archive -Path .\build\* -DestinationPath ".\build\Additional-Keys-0.6.x-1.0.0.zip"
